Update build-checks.yml
Reverting the changes. It is expected that the IAM package gets installed along with OS managed package upon scratch org creation. There is a package dependency on HealthCloud. It appears that Health Cloud installation key is not valid, so we would need to send a request to Salesforce to get a new package installation link.
Workaround: Proceed with post deployment steps to create the configuration of the IAM in Production.